ABC-77: expunge global mm_license from admin_console components
This commit takes a different tack than others on account of the heavy
use of the license object throughout the pages and the non-connectedness
of most components. Instead of denormalizing the individual properties
of the license per components, the `license` object itself is passed
down in most cases. Unfortunately, this means the `=== 'true'` checks
remain littered throughout, but those will be removed at the same time
we cut over all the connectors.